+++ kafka-websocket-consumer/src/main/java/com/sonvu/websocket/config/ApplicationBean.java
@@ -1,0 +1,14 @@
+package com.sonvu.websocket.config;
+
+
+import lombok.Data;
+
+@Data
+public class ApplicationBean {
+
+    public ApplicationBean() {
+
+    }
+    private String data;
+
+}
