+++ kafka-websocket-consumer/src/main/java/com/sonvu/websocket/domain/Message.java
@@ -1,0 +1,15 @@
+package com.sonvu.websocket.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Message {
+
+    private Long id;
+    private String item;
+    private Double amount;
+}
