+++ kafka-websocket-consumer/src/main/java/com/sonvu/websocket/domain/MessageContent.java
@@ -1,0 +1,11 @@
+package com.sonvu.websocket.domain;
+
+import lombok.Data;
+
+@Data
+public class MessageContent {
+
+    public String item = "";
+    public Double amount = null;
+}
+
