--- kafka-websocket-consumer/src/main/java/son/vu/websocket/config/ApplicationBean.java
@@ -1,14 +1,0 @@
-package son.vu.websocket.config;
-
-
-import lombok.Data;
-
-@Data
-public class ApplicationBean {
-
-    public ApplicationBean() {
-
-    }
-    private String data;
-
-}
