--- kafka-websocket-consumer/src/main/java/son/vu/websocket/config/WebSocketConfig.java
@@ -1,85 +1,0 @@
-package son.vu.websocket.config;
-
-import io.confluent.kafka.serializers.KafkaAvroDeserializer;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-import son.vu.avro.domain.SaleReport;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- *
- */
-@EnableKafka
-@Configuration
-@EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-    @Value("${spring.kafka.consumer.bootstrap-servers}")
-    public String bootstrapServers;
-
-    @Bean
-    public ConsumerFactory<String, SaleReport> consumerFactory() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, "group-id");
-
-        props.put("auto.commit.enable", "false");
-        props.put("auto.offset.reset", "earliest");
-        // avro part (deserializer)
-        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, AvroDeserializer.class);
-        props.put("schema.registry.url", "http://127.0.0.1:8081");
-        props.put("specific.avro.reader", "true");
-        return new DefaultKafkaConsumerFactory<>(props, new StringDeserializer(), new AvroDeserializer<>(SaleReport.class));
-    }
-    @Bean
-    public ConcurrentKafkaListenerContainerFactory<String, SaleReport> kafkaListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, SaleReport>
-                factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(consumerFactory());
-        return factory;
-    }
-
-    @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
-    @Bean
-    public ApplicationBean getApplicationBean() {
-        return new ApplicationBean();
-    }
-
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-
-        registry.addEndpoint("/socket")
-                .setAllowedOriginPatterns("*")
-                .withSockJS();
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        registry.setApplicationDestinationPrefixes("/app");
-        registry.enableSimpleBroker("/message", "/passport");   // Enables a simple in-memory broker
-
-    }
-}
