--- kafka-websocket-consumer/src/main/java/son/vu/websocket/controller/WebSocketController.java
@@ -1,38 +1,0 @@
-package son.vu.websocket.controller;
-
-import son.vu.websocket.config.ApplicationBean;
-import son.vu.websocket.model.Passport;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class WebSocketController {
-
-    private final SimpMessagingTemplate template;
-
-    final ApplicationBean applicationBean;
-
-    @Autowired
-    WebSocketController(SimpMessagingTemplate template, ApplicationBean applicationBean){
-        this.template = template;
-        this.applicationBean = applicationBean;
-    }
-
-    @MessageMapping("/send/message")
-    public void sendMessage(String message) {
-        if(message.equals("hello")) {
-            this.template.convertAndSend("/message",  applicationBean.getData());
-        } else {
-            this.template.convertAndSend("/message",  message);
-        }
-
-    }
-
-    @MessageMapping("/send/passport")
-    public void sendPassport( Passport passport){
-        System.out.println(passport);
-        this.template.convertAndSend("/passport",  passport);
-    }
-}
