--- kafka-websocket-consumer/src/main/java/son/vu/websocket/domain/Message.java
@@ -1,15 +1,0 @@
-package son.vu.websocket.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Message {
-
-    private Long id;
-    private String item;
-    private Double amount;
-}
