--- kafka-websocket-consumer/src/main/java/son/vu/websocket/domain/MessageContent.java
@@ -1,11 +1,0 @@
-package son.vu.websocket.domain;
-
-import lombok.Data;
-
-@Data
-public class MessageContent {
-
-    public String item = "";
-    public Double amount = null;
-}
-
