--- kafka-websocket-consumer/src/main/java/son/vu/websocket/domain/dto/MessageDto.java
@@ -1,16 +1,0 @@
-package son.vu.websocket.domain.dto;
-
-import lombok.Data;
-import lombok.Value;
-
-import java.util.List;
-
-@Data
-@Value
-public class MessageDto {
-
-    List<List<String>> products;
-
-    String item;
-    Double amount;
-}
