--- kafka-websocket-consumer/src/main/resources/avro/customer-v1.avsc
@@ -1,16 +1,0 @@
-{
-     "type": "record",
-     "namespace": "son.vu.avro.domain",
-     "name": "Customer",
-     "version": "1",
-     "fields": [
-       { "name": "first_name", "type": "string", "doc": "First Name of Customer" },
-       { "name": "last_name", "type": "string", "doc": "Last Name of Customer" },
-       { "name": "age", "type": "int", "doc": "Age at the time of registration" },
-       { "name": "height", "type": "float", "doc": "Height at the time of registration in cm" },
-       { "name": "weight", "type": "float", "doc": "Weight at the time of registration in kg" },
-       { "name": "automated_email", "type": "boolean", "default ": true, "doc": "Field indicating if the user is enrolled in marketing emails" },
-       { "name": "phone_number", "type": ["null", "string"], "default": null, "doc": "optional phone number"},
-       { "name": "email", "type": "string", "default": "missing@example.com", "doc": "email address"}
-     ]
-}
