--- kafka-websocket-consumer/src/main/resources/avro/sale.avsc
@@ -1,13 +1,0 @@
-{
-     "type": "record",
-     "namespace": "son.vu.avro.domain",
-     "name": "SaleDetail",
-     "version": "1",
-     "fields": [
-       { "name": "sales_date", "type": "string", "doc": "Sale Date" },
-       { "name": "store_name", "type": "string", "doc": "Store Name" },
-       { "name": "product_name", "type": "string", "doc": "Product Name" },
-       { "name": "sales_units", "type": "int", "doc": "Sale Units" },
-       { "name": "sales_revenue", "type": "float", "doc": "Sale Revenue" }
-     ]
-}
