+++ kafka-websocket-producer/src/main/java/com/sonvu/producer/domain/MessageContent.java
@@ -1,0 +1,10 @@
+package com.sonvu.producer.domain;
+
+import lombok.Data;
+
+@Data
+public class MessageContent {
+
+    public String item = "";
+    public Double amount = null;
+}
