--- kafka-websocket-producer/src/main/java/son/vu/producer/domain/MessageContent.java
@@ -1,12 +1,0 @@
-package son.vu.producer.domain;
-
-import lombok.*;
-
-import java.util.List;
-
-@Data
-public class MessageContent {
-
-    public String item = "";
-    public Double amount = null;
-}
