--- kafka-websocket-producer/src/main/java/son/vu/producer/domain/SalesDetailBean.java
@@ -1,22 +1,0 @@
-package son.vu.producer.domain;
-
-import com.opencsv.bean.CsvBindByPosition;
-import lombok.Data;
-
-@Data
-public class SalesDetailBean {
-    @CsvBindByPosition(position = 0)
-    private String salesDate;
-
-    @CsvBindByPosition(position = 1)
-    private String storeName;
-
-    @CsvBindByPosition(position = 2)
-    private String productName;
-
-    @CsvBindByPosition(position = 3)
-    private String salesUnits;
-
-    @CsvBindByPosition(position = 4)
-    private String salesRevenue;
-}
