+++ kafka-websocket-producer/src/main/resources/avro/sale-report-v1.avsc
@@ -1,6 +1,6 @@
 {
   "type": "record",
-  "namespace": "son.vu.avro.domain",
+  "namespace": "com.sonvu.avro.domain",
   "name": "SaleReport",
   "fields": [
     {
