+++ kafka-websocket-producer/src/test/java/com/sonvu/producer/KafkaWebSocketProducerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sonvu.producer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KafkaWebSocketProducerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
