--- kafka-websocket-producer/src/test/java/son/vu/producer/KafkaWebSocketProducerApplicationTests.java
@@ -1,13 +1,0 @@
-package son.vu.producer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KafkaWebSocketProducerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
